popup_run_prompt: remove most references to mouse.screen
This commit removes most references to mouse.screen, as it wouldn't make
sense to have more than one prompts running at the same time (as all try
to grab the keyboard). This also fixes the bug where opening a
promptbox, moving the mouse to another screen and then closing the
promptbox results in the box staying on the screen.

Signed-off-by: Gregor Best <gbe@intepi.net>