Add a widget which shows the used space on a file system
Usage: obvious.fs_usage("/home") or something similar. The widget will display
the fs usage on the file system on which <path> resides.

This widget is possibly not linux specific since it uses /bin/df to get the
needed statistics. (Linux doesn't seem to export those through /proc or /sys)

Signed-off-by: Uli Schlachter <psychon@znc.in>
Signed-off-by: Gregor Best <gbe@intepi.net>