keymap_switch: Use spawn() instead of popen()
popen() returns a file handle which can be used to interact with the started
process. This file handle is only closed when it's explicitly closed (not
happening here) or when it is garbage collected.

Turns out that the lua garbage collector kicks in seldom enough for my awesome
instance to run into the fd limit (1024 open FDs is the maximum by default).

Fix this by using awful.util.spawn() instead of io.popen(). This will make the
output of setxkbmap visible as awesome's output, but besides that it should work
as good as the old code.

Signed-off-by: Uli Schlachter <psychon@znc.in>
Signed-off-by: Gregor Best <gbe@ring0.de>