Fix repeating nil error on double unregister
If you register the same function as a timer action twice, unregistering
it will fail, because the first unregister call will clear out the
reference to the function in the registry.  This change just makes it
so that unregistering a function twice doesn't cause a nil lookup error.

A better solution (probably) would be to make sure that a function can't
be registered twice, but I have a sinking feeling that obvious relies on
being able to do this