Minimize fluctuations in music marquee width
Before this change, the music widget, when in marquee mode, had an
annoying behavior - some songs' substrings would have different widths,
causing the width of the widget itself to fluctuate as the marquee
scrolled by.

This change doesn't stop fluctuation entirely, but it makes sure that
the widget only grows so that any fluctuation stops after the scrolling
has looped through the song title once.